Moved publicly accessable routes to public.py & repaired redirects this broke
