+++ safety-quiz/templates/layout.html
@@ -10,7 +10,7 @@
 
 {% block navbar %}
     <nav class="navbar navbar-expand-lg navbar-dark">
-        <a class="navbar-brand" href="{{ url_for('index') }}">Idea Shop Portal</a>
+        <a class="navbar-brand" href="{{ url_for('public.index') }}">Idea Shop Portal</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
             <span class="navbar-toggler-icon"></span>
         </button>
@@ -22,7 +22,7 @@
             {% if session['admin'] %}
                 <a class="nav-item nav-link" style="color: white" href="{{ url_for('reservation.view_reservations') }}" {% if not session['sid'] %} hidden="true" {% endif %}>View Reservations</a>
             {%  endif %}
-                <a class="nav-item nav-link" style="color: white" href="{{ url_for('shop_status') }}">Shop Status</a>
+                <a class="nav-item nav-link" style="color: white" href="{{ url_for('public.shop_status') }}">Shop Status</a>
                 <a class="nav-item nav-link" style="color: white" href="https://wiki.ideashop.iit.edu">Idea Shop Wiki</a>
             </div>
             <div class="navbar-nav ml-auto">
